fix: refresh nft list
