+++ packages/maskbook/src/social-network-adaptor/twitter.com/ui-provider.ts
@@ -29,7 +29,8 @@ import { currentSelectedIdentity } from '../../settings/settings'
 import { injectPostInspectorAtTwitter } from './injection/PostInspector'
 import { ProfileIdentifier } from '../../database/type'
 import { unreachable } from '@dimensiondev/kit'
-import { injectEnhancedProfile } from './injection/EnhancedProfileTab'
+import { injectEnhancedProfileTabAtTwitter } from './injection/EnhancedProfileTab'
+import { injectEnhancedProfileAtTwitter } from './injection/EnhancedProfile'
 
 const twitterUI: SocialNetworkUI.Definition = {
     ...twitterBase,
@@ -78,7 +79,8 @@ const twitterUI: SocialNetworkUI.Definition = {
         pageInspector: injectPageInspectorDefault(),
         postInspector: injectPostInspectorAtTwitter,
         setupPrompt: injectSetupPromptAtTwitter,
-        enhancedProfile: injectEnhancedProfile,
+        enhancedProfile: injectEnhancedProfileAtTwitter,
+        enhancedProfileTab: injectEnhancedProfileTabAtTwitter,
         newPostComposition: {
             start: injectPostBoxComposed,
             supportedInputTypes: {
