+++ packages/maskbook/src/utils/hooks/useLocationChange.ts
@@ -1,0 +1,8 @@
+import { useEffect } from 'react'
+
+export const useLocationChange = (handler: () => void) => {
+    useEffect(() => {
+        window.addEventListener('locationchange', handler)
+        return () => window.removeEventListener('locationchange', handler)
+    }, [handler])
+}
